requirements.txt
et-xmlfile==1.0.1
python-dateutil==2.8.1
pytz==2021.1
schedule==1.0.0
six==1.15.0
speedtest-cli==2.1.2
